#97 combine research and upgrades (i.e. merge actions & percepts)
