--- eisbw/src/main/java/eisbw/percepts/UpgradePercept.java
@@ -1,22 +1,0 @@
-package eisbw.percepts;
-
-import eis.iilang.Identifier;
-import eis.iilang.Percept;
-
-/**
- * @author Danny & Harm - The Upgrade percept.
- *
- */
-public class UpgradePercept extends Percept {
-	private static final long serialVersionUID = 1L;
-
-	/**
-	 * The UpgradePercept constructor.
-	 *
-	 * @param upgrade
-	 *            The name of the upgrade type.
-	 */
-	public UpgradePercept(String upgrade) {
-		super(Percepts.UPGRADING, new Identifier(upgrade));
-	}
-}
